Quick_Commerce/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
python-dotenv==1.0.0
pandas
sqlalchemy
pydantic
faker